Backend/api_gateway/app/requirements.txt
fastapi
uvicorn[standard]
httpx
PyJWT
python-multipart
slowapi
itsdangerous

# OpenTelemetry
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-httpx
